xcvxcv

+++ src/Program/Manager/CameraManager.cpp
@@ -25,7 +25,7 @@ namespace Chrivent {
 	int CameraManager::GetLastFrame() const {
 		if (!cameraAnim)
 			return 0;
-		return static_cast<int>((std::min)(cameraAnim->GetLastFrame(), static_cast<uint32_t>((std::numeric_limits<int>::max)())));
+		return (std::min)(cameraAnim->GetLastFrame(), static_cast<uint32_t>((std::numeric_limits<int>::max)()));
 	}
 
 	const std::vector<CameraAnimationKey>& CameraManager::GetAnimationKeys() const {
